Fixing some CPLEX parameters
git-svn-id: svn+ssh://external.hltcoe.jhu.edu/home/hltcoe/mgormley/public/repos/dep_parse_filtered/trunk@107 0408881f-cb4b-4750-8b71-c94bcb194269

+++ src/edu/jhu/hltcoe/ilp/IlpSolverFactory.java
@@ -8,9 +8,9 @@ public class IlpSolverFactory {
     
     private IlpSolverId ilpSolver;
     private int numThreads;
-    private int workMemMegs;
+    private double workMemMegs;
 
-    public IlpSolverFactory(IlpSolverId ilpSolver, int numThreads, int workMemMegs) {
+    public IlpSolverFactory(IlpSolverId ilpSolver, int numThreads, double workMemMegs) {
         this.ilpSolver = ilpSolver;
         this.numThreads = numThreads;
         this.workMemMegs = workMemMegs;
